requirements.txt
requests
phonenumbers
dnspython
python-whois
